backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
scikit-learn==1.5.0
fairlearn==0.10.0
google-genai>=1.0.0
firebase-admin==6.5.0
python-dotenv==1.0.1
requests==2.32.3
joblib==1.4.2
numpy==1.26.4
slowapi==0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
